refactor: minimize public surface of exo-sql-pg* crates
Internal modules in `exo-sql-pg`, `exo-sql-pg-connect`, `exo-sql-pg-schema`, and `exo-sql-pg-testing` are now private. Public types, functions, and constants are re-exported flat at each crate root, so consumers import from `exo_sql_pg_schema::Foo` instead of `exo_sql_pg_schema::some_module::Foo`. External call sites are updated accordingly.

- Remove the unused `test_helper` module from `exo-sql-pg-schema`.
- `exo_sql_pg_testing` now exposes only the `run` entry point.
