+++ crates/postgres-subsystem/postgres-core-builder/src/database_builder.rs
@@ -634,11 +634,8 @@ fn determine_column_type<'a>(
     if let Some(hint) = &field.type_hint {
         let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
         if let Some(explicit) = hint_ref.downcast_ref::<ExplicitTypeHint>() {
-            return exo_sql_pg_schema::column_spec::physical_column_type_from_string(
-                &explicit.dbtype,
-                &vec![],
-            )
-            .unwrap();
+            return exo_sql_pg_schema::physical_column_type_from_string(&explicit.dbtype, &vec![])
+                .unwrap();
         }
     }
