+++ crates/testing/src/lib.rs
@@ -18,7 +18,7 @@ use std::sync::Arc;
 use anyhow::{Context, Result, bail};
 use colored::Colorize;
 
-use exo_sql_pg_connect::testing::db::EphemeralDatabaseLauncher;
+use exo_sql_pg_connect::testing::EphemeralDatabaseLauncher;
 
 use model::TestSuite;
